Expense-API/requirements.txt
# Dependencies
aio-pika~=9.5
aiokafka~=0.12
alembic~=1.17
fastapi[standard]~=0.120
filetype~=1.2
minio~=7.2
openpyxl~=3.1
pandas~=2.3
phonenumbers~=9.0
pydantic~=2.12
pydantic-settings~=2.11
PyJWT[crypto]~=2.10
psycopg2-binary~=2.9
redis~=7.0
sqlmodel~=0.0

# Dev Dependencies
pre-commit~=4.3
